back/AI_API/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.3
    # via
    #   -r requirements.in
    #   aiohttp
aiohttp==3.10.10
    # via
    #   -r requirements.in
    #   langchain
aiosignal==1.3.1
    # via
    #   -r requirements.in
    #   aiohttp
annotated-types==0.7.0
    # via
    #   -r requirements.in
    #   pydantic
anyio==4.6.2.post1
    # via
    #   -r requirements.in
    #   httpx
    #   openai
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
async-timeout==4.0.3
    # via
    #   -r requirements.in
    #   aiohttp
    #   langchain
attrs==24.2.0
    # via
    #   -r requirements.in
    #   aiohttp
certifi==2024.8.30
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
    #   requests
chardet==5.2.0
    # via
    #   -r requirements.in
    #   reportlab
charset-normalizer==3.4.0
    # via
    #   -r requirements.in
    #   requests
colorama==0.4.6
    # via
    #   -r requirements.in
    #   tqdm
distro==1.9.0
    # via
    #   -r requirements.in
    #   openai
django==4.1.13
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djongo
djangorestframework==3.15.1
    # via -r requirements.in
djongo==1.3.6
    # via -r requirements.in
dnspython==2.7.0
    # via
    #   -r requirements.in
    #   pymongo
exceptiongroup==1.2.2
    # via
    #   -r requirements.in
    #   anyio
frozenlist==1.4.1
    # via
    #   -r requirements.in
    #   aiohttp
    #   aiosignal
greenlet==3.1.1
    # via
    #   -r requirements.in
    #   sqlalchemy
gunicorn==23.0.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   -r requirements.in
    #   httpcore
httpcore==1.0.6
    # via
    #   -r requirements.in
    #   httpx
httpx==0.27.2
    # via
    #   -r requirements.in
    #   langgraph-sdk
    #   langsmith
    #   openai
httpx-sse==0.4.0
    # via
    #   -r requirements.in
    #   langgraph-sdk
idna==3.10
    # via
    #   -r requirements.in
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.6.1
    # via
    #   -r requirements.in
    #   openai
jsonpatch==1.33
    # via
    #   -r requirements.in
    #   langchain-core
jsonpointer==3.0.0
    # via
    #   -r requirements.in
    #   jsonpatch
langchain==0.3.3
    # via -r requirements.in
langchain-core==0.3.12
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
langchain-openai==0.2.2
    # via -r requirements.in
langchain-text-splitters==0.3.0
    # via
    #   -r requirements.in
    #   langchain
langgraph==0.2.39
    # via -r requirements.in
langgraph-checkpoint==2.0.1
    # via
    #   -r requirements.in
    #   langgraph
langgraph-sdk==0.1.33
    # via
    #   -r requirements.in
    #   langgraph
langsmith==0.1.136
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
msgpack==1.1.0
    # via
    #   -r requirements.in
    #   langgraph-checkpoint
multidict==6.1.0
    # via
    #   -r requirements.in
    #   aiohttp
    #   yarl
numpy==1.26.4
    # via
    #   -r requirements.in
    #   langchain
openai==1.52.0
    # via
    #   -r requirements.in
    #   langchain-openai
orjson==3.10.7
    # via
    #   -r requirements.in
    #   langgraph-sdk
    #   langsmith
packaging==24.1
    # via
    #   -r requirements.in
    #   gunicorn
    #   langchain-core
pillow==11.0.0
    # via
    #   -r requirements.in
    #   reportlab
propcache==0.2.0
    # via
    #   -r requirements.in
    #   yarl
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.23.4
    # via
    #   -r requirements.in
    #   pydantic
pymongo==4.10.1
    # via
    #   -r requirements.in
    #   djongo
python-decouple==3.8
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
regex==2024.9.11
    # via
    #   -r requirements.in
    #   tiktoken
reportlab==4.2.5
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via
    #   -r requirements.in
    #   langsmith
sniffio==1.3.1
    # via
    #   -r requirements.in
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.36
    # via
    #   -r requirements.in
    #   langchain
sqlparse==0.2.4
    # via
    #   -r requirements.in
    #   django
    #   djongo
tenacity==8.5.0
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
tiktoken==0.8.0
    # via
    #   -r requirements.in
    #   langchain-openai
tqdm==4.66.5
    # via
    #   -r requirements.in
    #   openai
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   anyio
    #   asgiref
    #   langchain-core
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2024.2
    # via
    #   -r requirements.in
    #   django
urllib3==2.2.3
    # via
    #   -r requirements.in
    #   requests
yarl==1.15.4
    # via
    #   -r requirements.in
    #   aiohttp

django-cors-headers == 4.5.0